refactor: move common functionality to base host
